fix: three bug fixes across firmware and kernel driver
firmware/rp2350b/src/spi_protocol.c:
- Fix printf format mismatch in handle_cmd_audio_mic_gain: the ternary
  expression 'gain_db > 24 ? 24 : gain_db' promotes to signed int due
  to integer promotion rules, causing a %u format mismatch flagged by
  cppcheck. Cast to (unsigned int) with unsigned literal comparands.

firmware/rp2350b/src/main.c:
- Fix double-escaped newline in critical battery shutdown printf: the
  string literal contained \r\n (producing literal backslash chars on
  serial output) instead of \r\n (CRLF). This caused the critical
  battery warning message to appear garbled on the UART console.

software/linux-drivers/src/apex_bridge.c:
- Fix potential deadlock in apex_sg_engine_stop: cancel_work_sync was
  called while sg_lock was held. The SG work handler reads eng->state
  in a loop guarded conceptually by sg_lock; holding the mutex while
  blocking in cancel_work_sync can deadlock if the work cannot observe
  the IDLE state change. Release sg_lock before cancel_work_sync, then
  re-acquire for buffer teardown.
- Fix pm_runtime resource leak in apex_bridge_probe error path: if
  pm_runtime_get_sync fails after pm_runtime_enable, the goto target
  was err_destroy_class which did not call pm_runtime_disable, leaving
  the PM framework in the enabled state. Add err_disable_pm label to
  call pm_runtime_disable before falling through to err_destroy_class.

Static analysis: cppcheck confirmed the spi_protocol.c format mismatch.
Remaining cppcheck findings are unusedFunction (cross-TU references,
not real unused code) and missing system headers (expected on host).
No RF register values, calibration constants, or protocol magic values
were modified. No .github/workflows files created or modified.

+++ firmware/rp2350b/src/spi_protocol.c
@@ -975,7 +975,7 @@ static void handle_cmd_audio_mic_gain(const uint8_t *payload, uint16_t len)
                gain_db, ret);
     } else {
         printf("AUDIO_MIC_GAIN: PGA gain set to %u dB\r\n",
-               gain_db > 24 ? 24 : gain_db);
+               (unsigned int)(gain_db > 24u ? 24u : gain_db));
     }
 }
